backend-api/requirements.txt
# Flask==3.0.3
# Flask-SQLAlchemy==3.1.1
# Flask-Bcrypt==1.0.1
# Flask-JWT-Extended==4.6.0
# Flask-Cors==4.0.1
# python-dotenv==1.0.1
# psycopg2-binary==2.9.9
# gunicorn==22.0.0
# Click==8.1.7



Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.1
Flask-Limiter==3.5.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
Click==8.1.7
redis==5.0.1
requests==2.31.0
pandas==2.1.4
numpy==1.24.4
Pillow==10.0.1
python-dateutil==2.8.2